transaction nonce and block nonce double

+++ templates/withKibana/blocks.go
@@ -21,7 +21,7 @@ var Blocks = Object{
 	"mappings": Object{
 		"properties": Object{
 			"nonce": Object{
-				"type": "long",
+				"type": "double",
 			},
 			"timestamp": Object{
 				"type":   "date",
